requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.2.1
fastapi==0.121.1
fastjsonschema==2.21.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
langchain==1.0.5
langchain-core==1.0.4
langchain-openai==1.0.2
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
openai==2.7.2
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
# uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
xxhash==3.6.0
yarg==0.1.9
zstandard==0.25.0
